prepared an app only build : Moved around video tutorial files, removed an extra activity to select a currency, remodeld the landing screen
