+++ app/src/main/java/org/myoralvillage/cashcalculator/SplashActivity.java
@@ -3,10 +3,13 @@ package org.myoralvillage.cashcalculator;
 import android.content.Intent;
 import android.content.pm.ActivityInfo;
 import android.os.Bundle;
+import android.util.Log;
+import android.util.TypedValue;
 import android.view.Window;
 import android.view.WindowManager;
 import android.widget.Button;
 import android.widget.ImageView;
+import android.widget.LinearLayout;
 
 import androidx.appcompat.app.AppCompatActivity;
 
@@ -17,6 +20,7 @@ public class SplashActivity extends AppCompatActivity {
 
     String currencyName = null;
     boolean numericMode = false;
+    private static SettingService settingService;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -26,13 +30,70 @@ public class SplashActivity extends AppCompatActivity {
         this.getWindow().setFlags(WindowManager.LayoutParams.FLAG_FULLSCREEN,
                 WindowManager.LayoutParams.FLAG_FULLSCREEN);
         setContentView(R.layout.activity_splash);
-        flagSelectListener();
-        mainActivityButtonListener();
+
+        settingService = new SettingService(getApplicationContext(), getResources());
+        buildLayout();
+
+//        flagSelectListener();
+//        mainActivityButtonListener();
         //Removed the Cash-Numeric switch from the Home Screen
         //modeSwitchButtonListener();
     }
 
-    private void mainActivityButtonListener() {
+    public static SettingService getSettingService() {
+        return settingService;
+    }
+
+    private void buildLayout() {
+        new CurrencyService(getApplicationContext(), settingService.getDefaultOrder()).call(currencies
+                -> runOnUiThread(() -> {
+            LinearLayout view = findViewById(R.id.currencies);
+            int width = (int) TypedValue.applyDimension(
+                    TypedValue.COMPLEX_UNIT_DIP,
+                    230f,
+                    getResources().getDisplayMetrics()
+            );
+            int height = (int) TypedValue.applyDimension(
+                    TypedValue.COMPLEX_UNIT_DIP,
+                    140f,
+                    getResources().getDisplayMetrics()
+            );
+            int margin = (int) TypedValue.applyDimension(
+                    TypedValue.COMPLEX_UNIT_DIP,
+                    20f,
+                    getResources().getDisplayMetrics()
+            );
+
+            for (int i = 0; i < currencies.length; i++) {
+                String currency = currencies[i];
+                Button button = new Button(this);
+                LinearLayout.LayoutParams params =
+                        new LinearLayout.LayoutParams(width, height);
+                if(i == 0){
+                    params.setMargins(margin*5, 0, 0, 0);
+                }else if (i + 1 == currencies.length) {
+                    params.setMargins(margin, 0, margin, 0);
+                }else{
+                    params.setMargins(margin, 0, 0, 0);
+                }
+
+                button.setLayoutParams(params);
+                button.setBackgroundResource(CurrencyService.getCurrencyResource(currency));
+                button.setOnClickListener(e -> switchToMainActivity(currency));
+                view.addView(button);
+            }
+        }));
+    }
+
+    private void switchToMainActivity(String currencyCode) {
+        Intent intent = new Intent(this, MainActivity.class);
+        intent.putExtra("currencyCode", currencyCode);
+        intent.putExtra("numericMode", getIntent().getBooleanExtra("numericMode", false));
+        startActivity(intent);
+        finish();
+    }
+
+/*    private void mainActivityButtonListener() {
         ImageView setting = findViewById(R.id.main);
         setting.setOnClickListener(e -> switchToTutorial());
     }
@@ -63,7 +124,7 @@ public class SplashActivity extends AppCompatActivity {
         btnSelectCountry.setOnClickListener(e -> openCountrySelector());
     }
 
-    /*private void modeSwitchButtonListener() {
+    private void modeSwitchButtonListener() {
         Switch modeSwitch = (Switch) findViewById(R.id.mode_switch);
         modeSwitch.setOnCheckedChangeListener(new CompoundButton.OnCheckedChangeListener() {
             @Override
@@ -71,12 +132,14 @@ public class SplashActivity extends AppCompatActivity {
                 numericMode = isChecked;
             }
         });
-    }*/
+    }
 
     private void openCountrySelector() {
         Intent intent = new Intent(this, SettingActivity.class);
         intent.putExtra("numericMode", numericMode);
         startActivity(intent);
         finish();
-    }
+    }*/
+
+
 }
