+++ app/src/main/res/layout/activity_splash.xml
@@ -7,59 +7,73 @@
     android:background="@drawable/wooden_surface"
     tools:context=".SplashActivity">
 
-    <Button
-        android:id="@+id/btn_select_country"
-        android:layout_width="154dp"
-        android:layout_height="83dp"
-        android:layout_marginStart="16dp"
-        android:layout_marginLeft="16dp"
-        android:layout_marginTop="84dp"
-        android:background="@drawable/kes"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
+<!--    <Button-->
+<!--        android:id="@+id/btn_select_country"-->
+<!--        android:layout_width="154dp"-->
+<!--        android:layout_height="83dp"-->
+<!--        android:layout_marginStart="16dp"-->
+<!--        android:layout_marginLeft="16dp"-->
+<!--        android:layout_marginTop="84dp"-->
+<!--        android:background="@drawable/kes"-->
+<!--        app:layout_constraintStart_toStartOf="parent"-->
+<!--        app:layout_constraintTop_toTopOf="parent" />-->
+
+    <!--    <ImageView-->
+    <!--        android:id="@+id/main"-->
+    <!--        android:layout_width="150dp"-->
+    <!--        android:layout_height="129dp"-->
+    <!--        android:layout_marginBottom="16dp"-->
+    <!--        android:visibility="visible"-->
+    <!--        app:layout_constraintBottom_toBottomOf="parent"-->
+    <!--        app:layout_constraintEnd_toEndOf="parent"-->
+    <!--        app:srcCompat="@drawable/hand"/>-->
 
     <ImageView
-        android:id="@+id/imageView"
-        android:layout_width="300dp"
-        android:layout_height="75dp"
-        android:layout_marginTop="4dp"
+        android:id="@+id/logo_mov"
+        android:layout_width="210dp"
+        android:layout_height="65dp"
+        android:layout_marginLeft="20dp"
+        android:layout_marginTop="20dp"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent"
         app:srcCompat="@drawable/orallogo" />
 
     <ImageView
-        android:id="@+id/main"
-        android:layout_width="150dp"
-        android:layout_height="129dp"
-        android:layout_marginBottom="16dp"
-        android:visibility="visible"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:srcCompat="@drawable/hand" />
-
-    <ImageView
-        android:id="@+id/imageView2"
-        android:layout_width="249dp"
-        android:layout_height="151dp"
-        android:layout_marginStart="239dp"
-        android:layout_marginLeft="239dp"
-        android:layout_marginTop="128dp"
-        android:layout_marginEnd="207dp"
-        android:layout_marginRight="207dp"
-        android:layout_marginBottom="107dp"
+        android:id="@+id/logo4Share"
+        android:layout_width="140dp"
+        android:layout_height="100dp"
+        android:layout_marginBottom="10dp"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
         app:srcCompat="@drawable/fourshare" />
 
     <ImageView
-        android:id="@+id/imageView3"
-        android:layout_width="221dp"
-        android:layout_height="80dp"
-        android:layout_marginTop="16dp"
+        android:id="@+id/logo_orali"
+        android:layout_width="200dp"
+        android:layout_height="55dp"
+        android:layout_marginTop="30dp"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintTop_toTopOf="parent"
         app:srcCompat="@drawable/oralmobile" />
+
+    <HorizontalScrollView
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_gravity="center"
+        android:layout_margin="20dp"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent">
+
+        <LinearLayout
+            android:id="@+id/currencies"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:gravity="center"
+            android:orientation="horizontal"></LinearLayout>
+
+    </HorizontalScrollView>
 
  </androidx.constraintlayout.widget.ConstraintLayout>
